Extract CommentWebView to a separate file
